pyproject.toml
[project]
name = "invoice-reader"
version = "0.1.0"
description = "Manage your invoices."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.56",
    "fastapi[standard]>=0.115.4",
    "pydantic>=2.9.2",
    "python-multipart>=0.0.17",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
